requirements.txt
chromadb
discord.py
langchain-community
langchain-openai
PyMuPDF